first commit on this for a while... this is the start of access control. it's the puppetlabs actionpolicy code which would usually run inside mcollectived, heavily adapted.
by default it looks for policies in the SQL DB, but can also use the policy files like actionpolicy usually does. there will probably be a web UI for admins to manage these policies in the DB sometime, but it's not started yet.

there should be an indication given if you have access to run an action, before it is run.