fix initscript to correctly support thin.
The mcomaster status was not working if using thin.
This was happening because if called without pid argument,
thin was creating the pid file on /tmp/pids/thin.pid,
and the initscript expected it on MCOMASTER_HOME/tmp/pids/thin.*.pid.

Also the ports defined on /etc/sysconfig/mcomaster was not being respected
if using thin. It was missing the --port parameter for thin to listen on
MCOMASTER_PORT.

This commit fix thin port support and the thin pid location.