fix vagrant activemq install and mcollective version
The new version of activemq rpm package 5.9 does not require
the tanukiwrapper rpm, because it provides a new init script.

But the puppetlabs-activemq module, install one old init script
that require tanukiwrapper, this means that the activemq installation
is broken.

Also it was tanukiwrapper that required the java package, without it
the system is with activemq installed and without java, so it does not
work.

The puppetlabs-activemq must be patched upstream, until this happens
we are manually ensuring java and tanukiwrapper is installed.

this commit also updates the version of mcollective to support the
mcollective-client gem that is used by the rpm.

it is important to always update the mcollective package version after
changing the Gemfile mcollective version.